+++ flow-effector-watch.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+node_modules/.bin/flow status --include-warnings --color=always
+