+++ src/__test__/reducer.test.js
@@ -1,15 +1,14 @@
 // @flow
-
+/* eslint-disable no-unused-vars */
 import {
   createReducer,
-  createDomain,
   createRootDomain,
   type Domain,
   type Event,
   type Reducer,
-  rootDomain,
   joint,
   effectorMiddleware,
+  mill,
 } from '..'
 
 import {periodic} from 'most'
@@ -18,7 +17,7 @@ import {createStore, applyMiddleware, type Store, combineReducers} from 'redux'
 type State = {foo: 'bar'}
 
 test('reducer', () => {
-  const domain = rootDomain()
+  const domain = createRootDomain()
   const event1: Event<'ev', State> = domain.event('event1')
   const event2: Event<'ev2', State> = domain.event('event2')
   const reset: Event<void, State> = domain.event('reset')
@@ -40,7 +39,7 @@ test('reducer', () => {
 })
 
 test('joint', async() => {
-  const domain = rootDomain()
+  const domain = createRootDomain()
   const event1: Event<'ev', State> = domain.event('event1')
   const event2: Event<'ev2', State> = domain.event('event2')
   const event3: Event<'ev3', State> = domain.event('event3')
@@ -98,6 +97,72 @@ test('joint', async() => {
       ]
     }
   })
+})
+
+test('mill', async() => {
+  const domain = createRootDomain()
+  const event1: Event<'ev', State> = domain.event('event1')
+  const event2: Event<'ev2', State> = domain.event('event2')
+  const event3: Event<'ev3', State> = domain.event('event3')
+  const reset: Event<mixed, State> = domain.event('reset')
+
+  const reducerA: Reducer<{
+    actions: string[]
+  }> = createReducer({actions: ['bar']})
+    .on([event1, event2, event3], (state, payload) => ({
+      actions: [...state.actions, payload]
+    }))
+    .reset(reset)
+  const reducerB: Reducer<{
+    resets: number,
+  }> = createReducer({resets: 0})
+    .on(reset, (state, p) => ({
+      resets: state.resets + 1,
+    }))
+
+  const union = mill() // Mill
+    .and(reducerA) // Mill<A>
+    .and(reducerB) // Mill<A, B>
+    .joint((value1, {resets}) => ({
+      value1,
+      resets,
+    })) // Reducer<{value1: A, resets: number}>
+    .on(event3, (state, p) => ({
+      ...state,
+      resets: state.resets + 10,
+    }))
+    .on(event2, (e, event) => e)
+
+  const none: any = undefined
+  const state1 = union(none, event1('ev').raw())
+  expect(state1).toMatchSnapshot()
+  const state2 = union(state1, reset('').raw())
+  expect(state2).toMatchSnapshot()
+  const state3 = union(state2, event2('ev2').raw())
+  expect(state3).toMatchObject({
+    resets: 1,
+    value1: {
+      actions: [
+        'bar',
+        'ev2',
+      ]
+    }
+  })
+  const state4 = union(state3, event3('ev3').raw())
+  expect(state4).toMatchObject({
+    resets: 11,
+    value1: {
+      actions: [
+        'bar',
+        'ev2',
+        'ev3',
+      ]
+    }
+  })
+  const emptyUnion = mill()
+    .joint(() => 'default')
+  const state5 = emptyUnion(state4, event3('ev3').raw())
+  expect(state5).toBe('default')
 })
 
 test('falsey values should been handled correctly', async() => {