+++ src/effect/index.js
@@ -1,0 +1,3 @@
+//@flow
+
+export {EffectConstructor} from './effect'