+++ src/index.js
@@ -3,10 +3,27 @@
 export type {Event, Effect, Domain, Reducer, Store} from './index.h'
 export {
   createDomain,
-  rootDomain,
-  rootDomain as createRootDomain,
+  createRootDomain,
 } from './domain'
 export {effectorMiddleware} from './middleware'
 export {createReducer} from './reducer'
 export {joint} from './joint'
 export {createHaltAction} from './config'
+
+
+// Experimental API
+
+export {
+  createEvent,
+  createEffect,
+} from './default-domain'
+
+export type {
+  DomainAuto,
+  EffectAuto,
+  EventAuto,
+} from './default-domain'
+
+
+export {mill} from './mill'
+export type {MillType} from './mill'