+++ src/subject.js
@@ -1,4 +1,4 @@
 //@noflow
 /* eslint-disable no-unused-vars */
 
-export {async as subject} from 'most-subject'
+export {sync as subject} from 'most-subject'